+++ main.go
@@ -23,6 +23,7 @@ var (
 	fromEmail    string
 	subject      string
 	hookout      bool
+	insecure     bool
 	stdin        *os.File
 
 	emailSubjectTemplate = "Sensu Alert - {{.Entity.Name}}/{{.Check.Name}}: {{.Check.State}}"
@@ -43,6 +44,7 @@ func main() {
 	cmd.Flags().StringVarP(&toEmail, "toEmail", "t", "", "The 'to' email address")
 	cmd.Flags().StringVarP(&fromEmail, "fromEmail", "f", "", "The 'from' email address")
 	cmd.Flags().BoolVarP(&hookout, "hookout", "H", false, "Include output from check hook(s)")
+	cmd.Flags().BoolVarP(&insecure, "insecure", "i", false, "Use an insecure connection (unauthenticated on port 25)")
 
 	cmd.Execute()
 }
@@ -91,14 +93,15 @@ func checkArgs() error {
 	if len(toEmail) == 0 {
 		return errors.New("missing destination email address")
 	}
-	if len(smtpUsername) == 0 {
-		return errors.New("smtp username is empty")
-	}
-	if len(smtpPassword) == 0 {
-		return errors.New("smtp password is empty")
-	}
-	if len(smtpPassword) == 0 {
-		return errors.New("smtp password is empty")
+	if !insecure {
+		if len(smtpUsername) == 0 {
+			return errors.New("smtp username is empty")
+		}
+		if len(smtpPassword) == 0 {
+			return errors.New("smtp password is empty")
+		}
+	} else {
+		smtpPort = 25
 	}
 	if hookout {
 		emailBodyTemplate = "{{.Check.Output}}\n{{range .Check.Hooks}}Hook Name:  {{.Name}}\nHook Command:  {{.Command}}\n\n{{.Output}}\n\n{{end}}"
@@ -125,7 +128,28 @@ func sendEmail(event *types.Event) error {
 		"\r\n" +
 		body + "\r\n")
 
+	if insecure {
+		smtpconn, connErr := smtp.Dial(smtpAddress)
+		if connErr != nil {
+			return connErr
+		}
+		defer smtpconn.Close()
+		smtpconn.Mail(fromEmail)
+		smtpconn.Rcpt(toEmail)
+		smtpdata, dataErr := smtpconn.Data()
+		if dataErr != nil {
+			return dataErr
+		}
+		defer smtpdata.Close()
+		buf := bytes.NewBuffer(msg)
+		if _, dataErr := buf.WriteTo(smtpdata); dataErr != nil {
+			return dataErr
+		}
+
+		return nil
+	}
 	return smtp.SendMail(smtpAddress, smtp.PlainAuth("", smtpUsername, smtpPassword, smtpHost), fromEmail, []string{toEmail}, msg)
+
 }
 
 func resolveTemplate(templateValue string, event *types.Event) (string, error) {