+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added option to allow for including check hook output in the email body
+- Added option to allow for using insecure (port 25) email relays
 
 ## [0.0.2] - 2018-12-12