+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added option to allow for including check hook output in the email body
 - Added option to allow for using insecure (port 25) email relays
 - Added from address verification, RFC5322 From: header
+- Added body template file
 
 ## [0.0.2] - 2018-12-12