+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added option to allow for using insecure (port 25) email relays
 - Added from address verification, RFC5322 From: header
 - Added body template file
+- Added environment variables for SMTP username and password
 
 ## [0.0.2] - 2018-12-12