+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.0] - 2019-03-04
+
 ### Changed
 - Updated travis, goreleaser configurations.
 - Updated license.